Fix lỗi ngày 11-12-2021
========================== LỖI CHƯA SỬA ĐƯỢC =================================
- Trong suốt phân tán, giảng viên k đồng bộ kịp
- giảng viên hiện tên
Còn 1 lỗi concurrency violation ( coi video ) => Có thể lỗi xảy ra do có cột ngày tháng
https://stackoverflow.com/questions/3388100/concurrency-violation-the-updatecommand-affected-0-of-the-expected-1-records
=> Không rõ cách sửa, mà tôi không nghĩ là quan trọng dù sao cũng thấy ít khi xảy ra và cũng đã bắt lỗi
========================== LỖI ĐÃ SỬA =================================
- Tổng LT +TH >=1
- Học vị học hàm chuyên môn, trim() trước khi lưu
- Phái(nữ)
- số tc lt th k âm
- Không đóng học phí cho tương lai
- Lúc thêm sinh tạo thêm mật khẩu mặc định 123456
- Không cho xoá chính mình (Giảng viên)
Lỗi đăng ký lớp tín chỉ.
- Nút xoá đk disable khi đã lưu
- huỷ dkd thì chỉ cần tích vào thôi
Khi đóng xong vẫn ở chỗ cũ đóng chứ k đc về đầu tiên
- Tên mh, tên lớp: chữ cái có hoặc ko dấu, số, khoảng trắng, ko ký tự đặc biệt
- Lớp tín chỉ, khi thêm mới xong phải ở dòng mới thêm
lỗi lớp tín chỉ:
- Thêm mới, rồi phục hồi thì ko xoá cái thêm mới
- Hoàn tác không được.

+++ QLDSV_HTC/Forms/frmGiangVien.cs
@@ -110,19 +110,26 @@ namespace QLDSV_HTC.Forms
 
         private void btnXoa_ItemClick(object sender, DevExpress.XtraBars.ItemClickEventArgs e)
         {
-            string maGV = "";
             if (bdsLopTinChi.Count > 0)
             {
                 XtraMessageBox.Show("Không thể xóa giảng viên này vì đã có lớp tín chỉ do giảng viên này dạy!",
                     "Lỗi", MessageBoxButtons.OK, MessageBoxIcon.Error);
                 return;
             }
-            if (XtraMessageBox.Show("Bạn có chắc muốn xóa giảng viên này không?", "Xác nhận",
-                        MessageBoxButtons.OKCancel, MessageBoxIcon.Warning) != DialogResult.OK) return;
 
+            string maGV = "";
             try
             {
                 maGV = ((DataRowView)bdsGiangVien[bdsGiangVien.Position])["MAGV"].ToString();
+                if (string.Equals(maGV, Program.username, StringComparison.OrdinalIgnoreCase))
+                {
+                    XtraMessageBox.Show("Bạn không thể xóa chính mình!", "Lỗi",
+                       MessageBoxButtons.OK, MessageBoxIcon.Error);
+                    return;
+                }
+
+                if (XtraMessageBox.Show("Bạn có chắc muốn xóa giảng viên này không?", "Xác nhận",
+                        MessageBoxButtons.OKCancel, MessageBoxIcon.Warning) != DialogResult.OK) return;
                 cmd.Add(new DeleteAction(bdsGiangVien));
                 bdsGiangVien.RemoveCurrent();
                 gIANGVIENTableAdapter.Connection.ConnectionString = Program.connstr;
@@ -216,7 +223,7 @@ namespace QLDSV_HTC.Forms
                 txtHo.Focus();
                 return;
             }
-            if (!Regex.IsMatch(ho, Utils.CAC_TU_CO_DAU))
+            if (!Regex.IsMatch(ho, Utils.HO_NGUOI))
             {
                 XtraMessageBox.Show("Họ của giảng viên chỉ được bao gồm các chữ cái a-z, A-Z\n" +
                     "(có dấu hoặc không dấu) và khoảng trắng giữa các từ!",
@@ -232,7 +239,7 @@ namespace QLDSV_HTC.Forms
                 txtTen.Focus();
                 return;
             }
-            if (!Regex.IsMatch(ten, Utils.TU_CO_DAU))
+            if (!Regex.IsMatch(ten, Utils.TEN_NGUOI))
             {
                 XtraMessageBox.Show("Tên của giảng viên chỉ được bao gồm các chữ cái a-z, A-Z\n" +
                     "(có dấu hoặc không dấu) và chỉ gồm 1 từ duy nhất!",
@@ -267,6 +274,9 @@ namespace QLDSV_HTC.Forms
                 current["MAGV"] = maGV;
                 current["HO"] = ho;
                 current["TEN"] = ten;
+                current["HOCVI"] = txtHocVi.Text.Trim();
+                current["HOCHAM"] = txtHocHam.Text.Trim();
+                current["CHUYENMON"] = txtChuyenMon.Text.Trim();
                 gIANGVIENTableAdapter.Connection.ConnectionString = Program.connstr;
                 this.gIANGVIENTableAdapter.Update(this.DS.GIANGVIEN);
                 if (optionThem) cmd.Add(new InsertAction(bdsGiangVien));
